final components + file reorganization
