+++ pages/login.tsx
@@ -1,11 +1,14 @@
 import { NextPage } from "next";
 import React, { useState, useEffect } from "react";
 import styles from "../styles/Login.module.css";
-import { LoginPageMain, LoginNameInput, LoginPositionInput } from "../components/LoginViews";
+import LoginPageMain from "../components/Login/LoginPageMain";
+import LoginNameInput from "../components/Login/LoginNameInput";
+import LoginPositionInput from "../components/Login/LoginPositionInput";
+import CreatePassword from "../components/Login/CreatePassword";
 
 const Login: NextPage = () => {
     
-    return <LoginPositionInput />;
+    return <CreatePassword />;
 }
 
 export default Login;
